+++ Sources/Main/RAT/RAnalyticsRATTracker.swift
@@ -2,12 +2,6 @@ import Foundation
 import CoreTelephony
 import CoreLocation
 import UIKit
-#if canImport(RSDKUtils)
-import RSDKUtils
-#else // SPM version
-import RSDKUtilsMain
-import RLogger
-#endif
 
 // swiftlint:disable type_name
 public typealias RAnalyticsRATShouldDuplicateEventCompletion = (_ eventName: String, _ duplicateAccId: Int64) -> Bool